[ANDROSDK-688] Unify UserOrganisationUnitLink store and handler
